Extract Eq and Ord member generation
Amp-Thread-ID: https://ampcode.com/threads/T-019fb21f-ec77-74dc-a82d-157ef6f83d0c
